% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_seq.R
\name{clean_seq_influenza_B}
\alias{clean_seq_influenza_B}
\title{Clean influenza B sequences}
\usage{
clean_seq_influenza_B(seqs.obj, maxreltol = 0.005, verbose = FALSE)
}
\arguments{
\item{seqs.obj}{List of sequences as returned from \code{import_seqs()}}

\item{maxreltol}{Numeric. Maximum tolerance for the
relative proportion of missing/aberrant proteins in any given sequence.
Default = 0.005.}
}
\value{
A list of sequences cleaned according to influenza A specifications.
}
\description{
Clean influenza B sequences
}
