% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_seqs.R
\name{clean_strain_names_location_seq}
\alias{clean_strain_names_location_seq}
\title{Remove underscores and spaces from location of strain names
A/Hong Kong/... ==> A/HongKong/...
A/Hong_Kong/... ==> A/HongKong/...}
\usage{
clean_strain_names_location_seq(x)
}
\description{
Remove underscores and spaces from location of strain names
A/Hong Kong/... ==> A/HongKong/...
A/Hong_Kong/... ==> A/HongKong/...
}
