% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist.R
\name{dist_blosum}
\alias{dist_blosum}
\title{BLOSUM distance between two sequences.}
\usage{
dist_blosum(x, y, sites = NULL, logistic.scale = 15)
}
\arguments{
\item{x}{Character vector representing the amino acids sequence.}

\item{y}{Character vector representing the amino acids sequence.}

\item{logistic.scale}{Numeric. Scale parameter for the logistic function
when converting similarity into a distance.
Default \code{logistic.scale = 15}.}
}
\value{
Numeric. BLOSUM80 distance between \code{x} and \code{y}.
}
\description{
Calculate the genetic distance according to
the BLOSUM metric. Here, only the BLOSUM80 is implemented.
BLOSUM80 definition: https://www.ncbi.nlm.nih.gov/IEB/ToolBox/C_DOC/lxr/source/data/BLOSUM80
}
\examples{
x = c('A','P','P','L','E','S')
y = c('A','P','M','L','E','S')
dist_blosum(x,y,  sites = list(c(1:2), 4:6))
dist_blosum(x,y,  sites = NULL)
}
