% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist.R
\name{dist_hamming}
\alias{dist_hamming}
\title{Hamming distance between two sequences.}
\usage{
dist_hamming(x, y, sites)
}
\arguments{
\item{x}{Character vector. Sequence 1.}

\item{y}{Character vector. Sequence 2.}

\item{sites}{List of integer vectors. Positions where the differences are calculated.
For example,  \code{sites = list( c(1:5), c(300:900), 1234}.
If \code{sites=NULL}, then all sites are included in the distance calculation.}
}
\value{
Hamming distance at the selected sites.
}
\description{
Hamming distance between two sequences.
}
