% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist.R
\name{dist_matrix}
\alias{dist_matrix}
\title{Distance matrix for multiple sequences.}
\usage{
dist_matrix(seqs, sites = NULL, ncores = 1, dist.type = "hamming")
}
\arguments{
\item{seqs}{List of sequences.}

\item{sites}{List of integer vectors. Positions where the differences are calculated.
For example,  \code{sites = list( c(1:5), c(300:900), 1234}.
If \code{sites=NULL} (Default), then all sites are included in the distance calculation.}

\item{ncores}{Integer. Number of cores used for parallel computation.
Default \code{ncores=1} (i.e., no parallel computation). To use all available cores, enter \code{ncores=0}.}

\item{dist.type}{String. Type of genetic distance. Default \code{dist.type='hamming'}.
Options are \code{'hamming', 'blosum'}.}
}
\value{
A distance matrix.
}
\description{
Distance matrix for multiple sequences.
}
