% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_seq.R
\name{filter_index}
\alias{filter_index}
\title{Title}
\usage{
filter_index(sobj, idx)
}
\arguments{
\item{sobj}{List as returned by the function \code{import_seqs()}.}

\item{idx}{Integer vector representing the indices of the sequences selected.}
}
\value{
A list of sequence objects.
}
\description{
Title
}
