% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_seq.R
\name{filter_seq_length}
\alias{filter_seq_length}
\title{Filter sequences based on their length}
\usage{
filter_seq_length(x, length.min, length.max, verbose = FALSE)
}
\arguments{
\item{x}{Sequence list, as returned by the function \code{angedist::import_seqs()}}

\item{length.min}{Integer. Minimum length (inclusive).}

\item{length.max}{Integer. Maximum length (inclusive).}

\item{verbose}{Logical. Display additional information about the filtering process.}
}
\value{
A list of sequences that all comply to the length specifications.
}
\description{
Filter sequences based on their length
}
