% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_seq.R
\name{filter_start_subseq}
\alias{filter_start_subseq}
\title{Filter sequences that have defined starting proteins.}
\usage{
filter_start_subseq(x, subseq, verbose = FALSE)
}
\arguments{
\item{x}{List of sequences.}

\item{subseq}{String. Sequence of characters that must start the sequence. For example \code{subseq='MK'} (not \code{c('M','K')}).}

\item{verbose}{Logical. Display additional information about the filtering process.}
}
\value{
A list of sequences that all comply with the tolerance for missing/aberrant sata.
}
\description{
Filter sequences that have defined starting proteins.
}
