% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_seqs.R
\name{get_accession_number}
\alias{get_accession_number}
\title{Retrieve the accession/ID number from the FASTA header.}
\usage{
get_accession_number(x, seq.source)
}
\arguments{
\item{x}{String. FASTA header.}

\item{seq.source}{String. Source of sequencing data (e.g., GISAID).}
}
\value{
String for the accession/ID number
}
\description{
Retrieve the accession/ID number from the FASTA header.
}
