% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_seqs.R
\name{get_collection_date}
\alias{get_collection_date}
\title{Get the collection date from a FASTA header.}
\usage{
get_collection_date(h)
}
\arguments{
\item{h}{A character vector of FASTA headers.}
}
\value{
A list of collection dates.
}
\description{
Get the collection date from a FASTA header.
}
