% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_seqs.R
\name{import_seqs}
\alias{import_seqs}
\title{Read and import genetic sequences from files.}
\usage{
import_seqs(path, prms)
}
\arguments{
\item{path}{String. Path to the file storing the genetic sequences.}

\item{prms}{List. Parameters for importation.}
}
\value{
A \code{SeqFastaXXX} object as defined by the package \code{seqinr}.
}
\description{
Read and import genetic sequences from files.
}
