% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist.R
\name{locate_postion}
\alias{locate_postion}
\title{Locate the most frequent starting position
of sub-sequence among multiple sequences.}
\usage{
locate_postion(sobj, subseq)
}
\arguments{
\item{sobj}{Sequence object as returned by \code{import_seqs()}.}

\item{subseq}{String. Sub-sequence to locate.}
}
\value{
The most frequent location.
}
\description{
Locate the most frequent starting position
of sub-sequence among multiple sequences.
}
