% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_seqs.R
\name{pathogen_code}
\alias{pathogen_code}
\title{Regex pathogen code in FASTA header.}
\usage{
pathogen_code(p, seq.source)
}
\arguments{
\item{p}{String. Pathogen name.}

\item{seq.source}{String. Source of sequence data (e.g., GISAID)}
}
\value{
String for a REGEX search.
}
\description{
Return the pathogen code as written in the FASTA header.
}
