% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mds.R
\name{plot_mds}
\alias{plot_mds}
\title{Plot a MDS object}
\usage{
plot_mds(
  mdsobj,
  color_varname = NULL,
  highlight = FALSE,
  highlight.label = FALSE
)
}
\arguments{
\item{mdsobj}{MDS object as returned by the function \code{angedist::mds()}.}

\item{color_varname}{String. Name of the variable used for the color aesthetic.
For no color, \code{color_varname = NULL} (default).}

\item{highlight}{Logical. Highlight MDS points that are identified as being
highlighted by defining them through a variable called \code{highlight}
in the dataframe \code{mdsobj$df}.}

\item{highlight.label}{Logical. Label highlighted MDS points with text defined
by the variable \code{highlight.label} in the dataframe \code{mdsobj$df}.}
}
\value{
A \code{ggplot2} object.
}
\description{
Plot a MDS object
}
