% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_seq.R
\name{remove_duplicated}
\alias{remove_duplicated}
\title{Remove duplicated sequences.}
\usage{
remove_duplicated(x, verbose = FALSE)
}
\arguments{
\item{x}{List of sequences.}

\item{verbose}{Logical. Display additional information about the filtering process.}
}
\value{
List of non-duplicated sequences.
}
\description{
Remove duplicated sequences.
}
