% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_seq.R
\name{remove_lowquality_seqs}
\alias{remove_lowquality_seqs}
\title{Remove low-quality sequences.}
\usage{
remove_lowquality_seqs(
  x,
  maxreltol,
  unwanted = c("X", "x", "-"),
  verbose = FALSE
)
}
\arguments{
\item{x}{List of sequences.}

\item{maxreltol}{Numerical. Maximum relative tolerance.}

\item{unwanted}{Character vector of unwanted characters in the sequence. Default is set to \code{c('X','x','-')}.}

\item{verbose}{Logical. Display additional information about the filtering process.}
}
\value{
A list of sequences that all comply with the tolerance for missing/aberrant sata.
}
\description{
Remove low-quality sequences.
}
