% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsample.R
\name{subsample_date}
\alias{subsample_date}
\title{Sub-sample sequences based on their collection date}
\usage{
subsample_date(sobj, n.subsample, years.exception = NULL, verbose = FALSE)
}
\arguments{
\item{sobj}{List of sequences as returned by \code{import_seqs()}.}

\item{n.subsample}{Integer. Size of the subsample.}

\item{years.exception}{Integer vector. Exception years
(sequences from those years will not be sub-sampled).
If no exception, enter \code{NULL}.}

\item{verbose}{Logical.}
}
\value{
A list of sequences that have been sub-sampled.
}
\description{
Sub-sample sequences based on their collection date
}
