% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mds.R
\name{plot_mds}
\alias{plot_mds}
\title{Plot a MDS object}
\usage{
plot_mds(mdsobj, color_varname = NULL)
}
\arguments{
\item{mdsobj}{MDS object as returned by the function \code{angedist::mds()}.}

\item{color_varname}{String. Name of the variable used for the color aesthetic.
For no color, \code{color_varname = NULL} (default).}
}
\value{
a \code{ggplot2} object.
}
\description{
Plot a MDS object
}
